requirements copy.txt
absl-py==2.2.1
alembic==1.14.1
argcomplete==3.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
azure-core==1.32.0
azure-storage-blob==12.25.0
beautifulsoup4==4.13.3
blinker==1.8.2
boto3==1.37.26
botocore==1.37.26
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmake==3.31.6
colorama==0.4.6
colorlog==6.9.0
ConfigArgParse==1.7
contourpy==1.1.1
cryptography==44.0.2
cycler==0.12.1
databricks-sdk==0.50.0
distro==1.9.0
docker==7.1.0
docutils==0.20.1
easydict==1.13
empy==3.3.4
evo==1.31.0
fastrlock==0.8.3
filelock==3.16.1
Flask==3.0.3
fonttools==4.56.0
fsspec==2024.6.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==1.0.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.11.0
hdf5plugin==5.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
llvmlite==0.41.1
lz4==4.3.3
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
minio==7.2.10
mlflow==2.17.2
mlflow-skinny==2.17.2
mpmath==1.3.0
narwhals==1.35.0
natsort==8.4.0
netifaces==0.11.0
networkx==3.0
nose==1.3.7
numba==0.58.1
numexpr==2.8.6
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.11.0.86
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
optuna==4.3.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pip==24.3.1
plotly==6.0.1
plyfile==1.0.3
progress==1.6
protobuf==3.20.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.22.0
Pygments==2.19.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0
python-gnupg==0.5.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
s3transfer==0.11.4
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
setuptools==75.0.0
six==1.16.0
smmap==5.0.2
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tqdm==4.67.1
#triton==2.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
Werkzeug==3.0.6
wheel==0.45.1
wrapt==1.17.2
yacs==0.1.8
zipp==3.20.2
zstandard==0.23.0
